Fixed a bug regarding flag handling and mode issues
Fixed an almost certain problem in the honeyd_overload.c file where
a flag was being updated (mask off) and then the flag holder was
set to the flag. Either the inverted mask bitwise and should have
been omitted or the assignment should have been a bitwise or with
the flag.

Fixed a minor mode issue on the subsystems smtp file. The C file
shouldn't be executable.